Change space settings management into tab views
