+++ gui/app/components/section/wysiwyg/type-editor.js
@@ -23,10 +23,6 @@ export default Component.extend({
 		let page = this.get('page');
 		return `wysiwyg-editor-${page.id}`;
 	}),
-	toolbarId: computed('page', function () {
-		let page = this.get('page');
-		return `wysiwyg-editor-toolbar-${page.id}`;
-	}),
 
 	didReceiveAttrs() {
 		this._super(...arguments);
@@ -45,7 +41,6 @@ export default Component.extend({
 				gecko_spellcheck: false,
 				statusbar: false,
 				inline: true,
-				// fixed_toolbar_container: '#' + this.get('toolbarId'),
 				paste_data_images: true,
 				image_advtab: true,
 				image_caption: true,
