+++ gui/app/styles/layout/all.scss
@@ -2,4 +2,5 @@
 @import "layout-topbar.scss";
 @import "layout-sidebar.scss";
 @import "layout-footer.scss";
+@import "layout-content.scss";
