+++ gui/app/styles/layout/layout-content.scss
@@ -1,0 +1,32 @@
+.content-zone {
+	> .explainer-header {
+		color: $color-gray;
+		font-size: 1.5rem;
+		font-weight: 500;
+	}
+
+	> .explainer-text {
+		margin: 3px 0;
+		padding: 0;
+		color: $color-gray;
+		font-size: 1.1rem;
+	}
+
+	> .explainer-list {
+		margin: 5px 20px;
+		padding: 0;
+		color: $color-gray;
+
+		> li {
+			margin: 0;
+			padding: 0;
+			font-size: 1.1rem;
+			list-style: square;
+		}
+
+	}
+
+	> .explainer-gap {
+		margin-bottom: 3rem;
+	}
+}
