+++ gui/app/templates/components/customize/license-key.hbs
@@ -12,7 +12,7 @@
 <div class="view-customize">
 	<form class="mt-5 ">
 		<div class="form-group row">
-			<label for="smtp-host " class="col-sm-2 col-form-label ">Enterprise Edition License Key</label>
+			<label for="smtp-host " class="col-sm-2 col-form-label ">Enterprise Edition<br/> License (optional)</label>
 			<div class="col-sm-10 ">
 				{{textarea value=model.license rows="10" class=(if LicenseError 'form-control is-invalid' 'form-control')}}
 				<small class="form-text text-muted ">XML format</small>
