+++ gui/app/templates/components/document/block-editor.hbs
@@ -2,7 +2,7 @@
 	<div class="row">
 		<div class="col-12">
 			<div class="block-editor">
-				{{component editorType document=document folder=folder page=page meta=meta blockMode=true onCancel=(action 'onCancel') onAction=(action 'onAction')}}
+				{{component editorType document=document space=space page=page meta=meta blockMode=true onCancel=(action 'onCancel') onAction=(action 'onAction')}}
 			</div>
 		</div>
 	</div>
