+++ gui/app/templates/components/folder/settings-blocks.hbs
@@ -1,0 +1,18 @@
+<div class="content-zone">
+	<div class="explainer-header explainer-gap">Content blocks provide re-usable content that can be inserted into any document</div>
+	{{#each blocks as |block|}}
+		<div class="mb-5">
+			<h1>{{block.title}}</h1>
+			<p>{{block.excerpt}}</p>
+			<button type="button" class="btn btn-outline-success" onclick={{action 'onEdit' block.id}}>Edit</button>
+			<div class="button-gap" />
+			<button type="button" class="btn btn-outline-danger" onclick={{action 'onShowDeleteDialog' block.id}}>Delete</button>
+		</div>
+	{{/each}}
+</div>
+
+{{#if isSpaceAdmin}}
+    {{#ui/ui-dialog title="Delete Content Block" confirmCaption="Delete" buttonType="btn-danger" show=showDeleteDialog onAction=(action 'onDeleteBlock')}}
+        <p>Are you sure you want to delete this re-usable content block?</p>
+    {{/ui/ui-dialog}}
+{{/if}}
