+++ gui/app/templates/components/folder/settings-templates.hbs
@@ -1,0 +1,10 @@
+<div class="content-zone">
+	<div class="explainer-header explainer-gap">Content Templates provide the basis for new documentation</div>
+	{{#each templates as |item|}}
+		<div class="mb-5">
+			<h1>{{item.title}}</h1>
+			<p>{{item.description}}</p>
+			<button type="button" class="btn btn-outline-success" onclick={{action 'onOpenTemplate' item.id}}>Open</button>
+		</div>
+	{{/each}}
+</div>
