+++ gui/app/templates/components/section/wysiwyg/type-editor.hbs
@@ -1,7 +1,6 @@
 {{#section/base-editor-inline document=document folder=folder page=page
 	blockMode=blockMode contentLinkerButton=true onInsertLink=(action 'onInsertLink')
 	isDirty=(action 'isDirty') onCancel=(action 'onCancel') onAction=(action 'onAction')}}
-	<div id={{toolbarId}}></div>
 	<div id={{editorId}} class="mousetrap wysiwyg wysiwyg-editor">
 		{{{pageBody}}}
 	</div>
